Creates security group and cluster
This creates a security group that references the VPC and binds it to
the cluster. It also creates the cluster with a node group. The nodes and
any pods that are added to the cluster appear to be on the private subnet.

I added instructions for how to verify that the stuff is ending up
on the private subnet in the readme.

This just uses the default IAM role. We should probably create our
own IAM role, so that's next.