Shows how to create a role which can be used for RBAC
Also has code which allows you to use a pulumi secret and some comments about that.

The secret is our aws account id which is used in creating the AssumeRole.

We may not end up using a role which is created and bound to the cluster for RBAC
however. Although it may be nice to have a role which is tied directly to the cluster
so that not everyone in our account has access to all roles.