Adds the s3 bucket to the role
This does let the RBAC role see the s3 bucket however that's only part of the
problem. The role also needs to have permission to do all the things that
pulumi does through its magic so it also shows my attempt to do that.

However we have decided to move away from this and just use the role for
connecting to the cluster and that's it.