Fixes stack state issues
When I went to try to destroy the cluster for the weekend I got an error from
pulumi saying that a resource urn didn't exist. Essentially the stack state became
different from what was actually deployed. I was able to recover and made a
few changes to the code at the same time which were bugs and which may
have caused things to go badly. This commit fixes that.

I also removed package.json and package-lock.json files that were in the
folio root. I must have done npm install one level down and caused that.