Parses list of modules from github
This gets the list of modules from platform-complete based on a certain provided
release. It keeps our declarative approach adding a new concept of prepare which
can and is unit tested using mocha chai. It has a new class called FolioModule
which we can use to represent all the details about deployment for a given module.

I'm thinking that we can define an interface called FolioModule with a required
deploy function that all subclasses will implement which will allow for all modules
-- regardless of their deviation from the basics, to be deployed.