Modules are deploying
This now takes the list of modules from a file in the releases directory.
Have seen the core modules deploy and transition quickly into the Running state.

The only thing that bothers me is that if the modules deploy into an error state
any future run will deploy new items rather than replace the old. I was able to
manually delete the old with kubectl delete deployment but if a lot of modules
were in this state it would be very tedious to delete them all manually.

I think what I'd like to try next time is pulumi refresh which is apparently
what you're supposed to do when things get out of sync. See the help on it.