Issue 43: deploy okapi and folio modules via helm (#122)
GETTING MODULES FROM A FILE

This reads modules to deploy in from a file. This is effectively the same file
from platform-complete from a given release branch, but will allow us to make
tweaks to that file should we want to create a build with  perhaps other modules.

BREAKING OKAPI AND OTHER SPECIAL MODULES OUT

This breaks okapi deployment out separately since it really is its own animal
and needs its own ingress LoadBalancer. This creates the load balancer and has
been verified to accept traffic at the external dns that amazon assigns to the
load balancer service.

I think we will want to do the same type of thing for other services that
require their own external access, such as edge-connexion.

This code also makes any modules dependent on the okapi deployment
working. Perhaps not strictly necessary, although probably a good thing since
if okapi deployment goes wrong, other modules will likely too.

KNOWN ISSUES WITH DEPLOYING HELM RELEASES

This now takes the list of modules from a file in the releases directory.
Have seen the core modules deploy and transition quickly into the Running state.

The only thing that bothers me is that if the modules deploy into an error state
any future run will deploy new items rather than replace the old. I was able to
manually delete the old with kubectl delete deployment but if a lot of modules
were in this state it would be very tedious to delete them all manually.

I think what I'd like to try next time is pulumi refresh which is apparently
what you're supposed to do when things get out of sync. See the help on it.

Co-authored-by: James Fuller <james.fuller.2008@gmail.com>