Added FolioDeployment class
This adds a new class which refactors a good bit of the code to define
a way of representing the whole folio deployment. We needed a way to represent
things like the tenant in lower levels of the code, so specifying the tenant
and other variables once at the top seemed like the way to go.

This unforunately doesn't successfully handle module registration. The post
job is running, which is good, I'm having trouble with what folio-helm expects
for resource names and what pulumi creates, with its random suffix.