Modules in tenant are registering
Requests to http://okapi:9130/_/proxy/tenants/cubl/modules are finally returning
the right results (the modules are being registered to the tenant). This means the
postJob hook is working in the helm chart. Setting up the db is still sketchy.
The create-db.sh is needed, but unclear if all of the commands are needed.

Lost time figuring out how to delete a pv. Adding to the readme about that.