Create database is fixed
There were a few problems with it. The container image it was pointing to
I don't think was the latest build. It was doing something with rds_superuser
for example. I rebuild the container and pushed it ghcr. Need to add execute
perms to shell script.

It is also idempotent. I have run it more than once against an existing database
and pulumi doesn't complain.