Tenant registration through postJob working
Seeing it work. Lots of refactoring to make steps dependent
on each other. What I'm not seeing work yet is bootstrapping the
supertenant. I added code for that, however I'm not able to register
mod-users-bl and I see no helpful messages in okapi. Registration just
to the tenant just fails silently. I'm assuming this is because it has
module dependencies.

So I suppose the next step is to enable more modules and see if I can
get the bootstrap super to run.