Superuser is creating; job is idempotent
There were some challenges in making the boostrap-superuser
container idempotent. The first was that it simply isn't written that
way. Instead it puts the burden on the operator to know what flags it
should be run with. Otherwise, running it twice puts the cluster in an
unstable state and certain modules have to be re-installed to fix it.

My workaround was to respect its design, and store a variable in a somewhat
revised deployment yaml called createSuperuser. When this is true
it will create it. Otherwise it will attempt to apply all the permissions
to the su. This needs to be run anytime the modules change which isn't
really captured in the "bootstrap-superuser" name. But it does so we do.