Removing writing to the deployment config
Writing to the deployment config to set a the value of createSuperuser
was a bad idea because -- of course -- it strips comments and having comments
was a motivation for going with yaml.

This commit also adds some thoughts on how we work with jobs to the readme.
My motivation was to have the createSuperuser job run every time pulumi up
runs, but pulumi doesn't really have a construct for this outside of
destroy --target which may be what we want. Either way this is probably ready
to merge since that would happen outside of typescript and be done on the
command line.