Attempts at solving timeouts
Noting is working yet but this has various attempts at fixing the timeout
error I continue to get when trying to register modules. The first fix was
to increase the number of pg connections to 500. The second attempt was to set
an environment variable DB_QUERYTIMEOUT. Added comments to index with thoughts
on that var and why it isn't reaching the pods.