Moves cluster and namespace out of FolioDeployment
Also cleans up naming such that all resource names are passed in as params in index.ts.
Including cluster and namespace in FolioDeployment wasn't contributing to the problem
of the the create-database job not being able to connect to the cluster. So far the only
way I can make this job succeed is to export the kubeconfig for the newly created
cluster. Once it is exported, it works, which implies that pulumi k8s jobs have this
dependency. So either I need to do that export progromatically or we we need break
things up and rely on the operator do do that, which doesn't seem right.

In any case I'm now discovering that we have the option of initContainers which could
easily replace all of our jobs.