This appears to solve the module registration inconsistencies
The solution seems to be to be to run the module registration jobs as
initContainers. These are a special kind of kubernetes container that run
before the pod's application-level container(s). The crucial feature
is that each init container waits on the prior container completing
successfully, which is exactly what we need in the case of
registering modules.

In my single test, I saw 53 modules get registered successfully the
first time. This has never happened with the previous tests and the previous
code.

This commit also fixes the issue of the create-database job failing on
the "first run". Actually it wasn't the first run at all but rather that
a provider wasn't being specified for the job, so there was no kubeconfig
to bind.

The last step isn't completely error free. For some reason I'm getting back
a lot of messages saying that containers are in an "incomplete status".
Yet, 53 modules have been registered.