Issue 135 bootstrap superuser (#138)
This combines module tenant registration and superuser bootstrapping into the same job. The critical piece that makes module registration work is the module registration containers run as init containers to the superuser creation job. This ensures that each module registration container runs and completes before the next one runs. This also has a number of changes that were necessary to make end-to-end runs of the stack actually work reliably, including:
* Converting all helm operations to Releases
* Ensuring that dependsOn is always correct
* Ensuring that every operation has a provider so that a local kubeconfig isn't necessary

Finally this add the createsSuperuser construct in the config yaml. See the documentation of that in the code for more information.