This has pulumi creating RDS Aurora
It took some doc digging but this is how Pulumi does it. I haven't
finished but I really think we have a good chance of making this all
be controlled out of pulumi and going straight to RDS, and bypassing
the in-cluster db. This has been seen to be idempotent, not doing an
update/replace once RDS is there.

The commented out code shows how we can do everything that we currently
have in create-db.sh with Pulumi control. If we can't install modules
against this I'll eat my shoe.

Actually if that happens we should fix the module, not stick with in cluster
pg. I think an alternative is to just stay with in cluster pg like TAMU
is doing. But after spending a day wading into this I'm pretty convinced
that migrating from in cluster to RDS is not great.

Besides the extra hassle, I think it sets us up for failure in a few ways.
If installs only work in-cluster, upgrades are likely to be the same.
Are we going to migrate back to in-cluster for upgrades? Blah,
I say no.