Fixes issues with database permissions
Seeing most modules register without a problem. So the generall approach is
working in all but a few cases. For some reason mod-pubsub is erroring out and can't
find the relation pubsub_config.event_descriptor. I think this might be because
I installted the modules _before_ getting the permissions right.

In any case most of the fixes involved adding new stuff to create_db.sh because
RDS is subtly different from in cluster postgres.