Issue 29 provision rds instance to migrate (#141)
This bypasses in-cluster postgres completely, and stands up the cluster using RDS aurora. It has been tested all the way through module registration and superuser creation and it has been seen to work. The number of availability zones in the test between the VPC and RDS are symmetrical, 3 and 3, which appears to improve things.