LoadBalancers are working; need to register ui mods
This code is working as far as the eks load balancers are concerned. It also
adds a script json2yaml.ts for generating the modules array. The current and
I think very likely correct hypothesis is that the reason the mods aren't
showing up in the stripes UI is that we're not registering the frontend mods
to okapi. These are items from install.json that have the "folio_" prefix.

These items have to be exluded from module install, but included in the
module registration. This code now handles both.