This registers the front end modules in the folio-helm way
But it still doesn't work, in that no modules appear in UI. I had too much faith
in the folio-helm project. It grabs items from the stripes.json file, and those items
have no versions, which means it is taking the version from latest. So now there is
a complete mismatch between our registered front end versions and the flower
release versions.

So the solution is likely to disable these versions and enable the right
versions, which is a real drag.

But I'm still wodering, even with these modules registed (the wrong versions),
why nothing in the UI still?