Apps are now appearing in the front end
The problem was we were not registering front-end mods, and taking
care to install the right versions (from our release). Folio helm has a
way of registering the front end mods in one go. The only problem is that
it grabs the latest snapshot versions, which is not what we want.

This fixes that.