Inventory front end reg fix and de-coupling mod dep descriptor pushing from mod reg
Inventory front-end wasn't registering. This should fix that. Was missing s3 secrets
that folio-helm wanted. Also this takes steps to decouple module registration from
deployment descriptors. The consensus is mod reg should be done in one request,
whereas dd push, needs to be done in separate requests. We continue using a job
but this time, just pass in an unordered list of containers. This appears to work
well. Okapi doesn't complain, and we can still maintain the list of modules and
versions outside of the containers, controlling it through pulumi.

Now just need to test doing mod reg in one request with true/false for ref data/
sample data.