register-modules.sh is working with ref data true
This proves that the problem with my previous approach was running each
registration job as a separate request. While this should work, it simply
doesn't in all combinations of loading ref data etc. The theory is that
when the request is sent as list of modules, okapi handles awaiting each
job completing, whereas letting k8s wait on each script call is less
reliable.