Scratch environment is creating
Ran into a new error which had to do running out of file descriptors. Added note to
README.md as a workaround. Also, the pg cluster name needed to be modified to include
the stack which in this case is 'scratch'. It may be that we have other non-prod
stacks in the future but for now prod doesn't have the stack name in the cluster name
but scratch does.