Defines a ClusterEndpoint to handle switching between db clusters
This is in support of #237. The cluster endpoint provides persistent way
to refer to RDS db clusters, making it easy to switch between them by changing
the cluster id, which is now persisted in the pulumi config.

There is still some work to do however. What I stll need to do is make
the stack aware of a first run vs db cluster re-assignment. This can be
controlled by the new config value. If the config value exists, then it's
a cluster re-assignment (from a backup for example), if the config value doesn't
exist, then it is a first run, and a cluster should be created.