Creating or swapping a db cluster is now fully supported
The workflow is as follows:
1. Standing up a new cluster for the first time creates and binds
a fresh db cluster.
2. Setting the new configuration key for the cluster identifier
causes that cluster to be used and the initial cluster to be
deleted.

I think this is what we want. It supports initial testing
of a new deployment, with a new db and with sample data potentially.
And then we can use aurora to clone and swap in and out clusters
as needed for back up and restore in production or to populate
a deployment for testing.