Metrics going to CloudWatch now
This is an alternative to prometheus, which I think is better. It required adding a managed
policy arn for the cluster to be able to talk to CloudWatch. The CloudWatch UI is more usable
than prometheus. And I think we're going to want to take advantage of CloudWatch for other
things as well as for its alerting features.