Changes for aws-native
This compiles with `pulumi up` but required a lot of project config  changes.

Needed to call await on a top level function to resolve the result to something
other than a promise. This does that but needed a lot of changes to the TS/node
engine. That seems to be working. Background:

https://github.com/pulumi/pulumi/issues/5161