More search config; rolling back top-level await
The top-level await config changes were breaking the tests I had. Was able
to pass in the Promise wrapped Output by paying closer attention to the type
so I rolled back the top-level await idea.

Added more config for search including the secrets for the dashboard.
Next is installing the dashboard via helm.