Able to reach opensearch
This works but it may not be optimial from a security perspective.
In order to submit unsigned requests (what we want) it looks like we need
to forgo putting it in vpc, unless we modify the mod-search helm chart
which requires a un and pw.

This implemnts an OS un and pw (fined grained access control) and using
said I am able to authenticate using them outside the cluster.

I may still be able to lock it down to within the cluster with an IP
address filter, still going with non-vpc.