Switches to aws rather than aws native for opensearch
The support for opensearch in the aws classic namespace was just
recently merged. This has the advantage of making updates to the search
resource actually work (in native, you had to delete and recreate the
search resource each time, which was very very slow). This allows for
snappy in place updates which is more or less how all of our other
resources work.