Shows how to create a domain in the vpc with aws namespace
But the dealbreaker for us is that basic auth at lest requires a signed
request, whereas non-vpc does not.

If at some point we modify the folio-helm mod-search chart, making
un and pw optional we could go back to testing this, but for now it
is fine to forge ahead with non-vpc options.